requirement.txt
pandas
OTXv2
python-Wappalyzer
python-whois








